Move dashboard route to PageRouter as that class can interact with Laravel
