+++ packages/realtime-compiler/src/Routing/PageRouter.php
@@ -9,6 +9,7 @@ use Hyde\Pages\Concerns\HydePage;
 use Hyde\RealtimeCompiler\Concerns\InteractsWithLaravel;
 use Hyde\RealtimeCompiler\Concerns\SendsErrorResponses;
 use Hyde\RealtimeCompiler\Http\DashboardController;
+use Hyde\RealtimeCompiler\Http\HtmlResponse;
 use Hyde\Support\Models\Route;
 
 /**
@@ -29,6 +30,12 @@ class PageRouter
 
     protected function handlePageRequest(): Response
     {
+        if ($this->request->path === '/dashboard' && $this->canRenderDashboard()) {
+            return new HtmlResponse(200, 'OK', [
+                'body' => (new DashboardController())->show(),
+            ]);
+        }
+
         $html = $this->getHtml(Route::getOrFail($this->normalizePath($this->request->path))->getPage());
 
         return (new Response(200, 'OK', [
@@ -61,6 +68,12 @@ class PageRouter
         }
 
         return file_get_contents((new StaticPageBuilder($page))->__invoke());
+    }
+
+    protected function canRenderDashboard(): bool
+    {
+        // If a dashboard page file exists, we just continue with the default handling
+        return ! file_exists(BASE_PATH.'/_pages/dashboard.blade.php') && ! file_exists(BASE_PATH.'/_pages/dashboard.md');
     }
 
     public static function handle(Request $request): Response
